Make Visitor non-static in DOM driver
Visitor had a static reference to itself that was used by the DOM
classes. This was under the assumption that only one Visitor is run in a
JVM, which is invalidated when running batch DOM driver. This commit
refactors DOM classes so that a Visitor object is passed around instead
of a static reference.