Implement new cost metric for typed expressions
In this commit, we tune the enumerator to search for all expressions of
a target type within the search space, and give more priority to ones
that use most of the user-defined variables. For ones that use the same
number of user-defined variables, we sort them by simplicity (as usual).